Renamed to OptimizedSchemaUpdate
Ran code clean
